smart-meter-stromzaehler/requirements.txt
python-dotenv
flask
pyjwt
requests
cryptography
